requirements.txt
absl-py==2.2.2
attrs==25.3.0
blinker==1.9.0
brax==0.12.3
chex==0.1.89
click==8.2.0
cloudpickle==3.1.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
distrax==0.1.5
dm-env==1.6
dm-tree==0.1.9
dotmap==1.3.30
etils==1.12.2
evosax==0.2.0
Farama-Notifications==0.0.4
Flask==3.1.1
flask-cors==5.0.1
flax==0.10.6
fonttools==4.58.0
fsspec==2025.3.2
gast==0.6.0
glfw==2.9.0
grpcio==1.71.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.1.1
gymnax==0.0.8
humanize==4.12.3
importlib_resources==6.5.2
itsdangerous==2.2.0
jaxopt==0.8.5
Jinja2==3.1.6
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
ml_collections==1.1.0
ml_dtypes==0.5.1
msgpack==1.1.0
mujoco==3.3.2
mujoco-mjx==3.3.2
nest-asyncio==1.6.0
numpy==2.2.5
opt_einsum==3.4.0
optax==0.2.4
orbax-checkpoint==0.11.13
packaging==25.0
pandas==2.2.3
pillow==11.2.1
protobuf==6.31.0
pyglet==2.1.6
Pygments==2.19.1
PyOpenGL==3.1.9
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytinyrenderer==0.0.14
pytz==2025.2
PyYAML==6.0.2
rejax==0.1.1
rich==14.0.0
scipy==1.15.3
seaborn==0.13.2
simplejson==3.20.1
six==1.17.0
tensorboardX==2.6.2.2
tensorflow-probability==0.25.0
tensorstore==0.1.75
toolz==1.0.0
tqdm==4.67.1
treescope==0.1.9
trimesh==4.6.9
typing_extensions==4.13.2
tzdata==2025.2
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
